Reverted to 3.5.2
