+++ src/main/java/org/mule/module/magento/MagentoCloudConnector.java
@@ -8,22 +8,63 @@
 
 package org.mule.module.magento;
 
-import com.magento.api.*;
-import org.mule.api.annotations.*;
+import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.apache.log4j.Priority;
+import org.mule.api.ConnectionException;
+import org.mule.api.ConnectionExceptionCode;
+import org.mule.api.annotations.Connect;
+import org.mule.api.annotations.ConnectionIdentifier;
+import org.mule.api.annotations.Connector;
+import org.mule.api.annotations.Disconnect;
+import org.mule.api.annotations.MetaDataKeyRetriever;
+import org.mule.api.annotations.MetaDataRetriever;
+import org.mule.api.annotations.MetaDataSwitch;
+import org.mule.api.annotations.Processor;
+import org.mule.api.annotations.QueryTranslator;
+import org.mule.api.annotations.ValidateConnection;
+import org.mule.api.annotations.display.Password;
 import org.mule.api.annotations.display.Placement;
+import org.mule.api.annotations.param.ConnectionKey;
 import org.mule.api.annotations.param.Default;
 import org.mule.api.annotations.param.Optional;
 import org.mule.api.annotations.param.Payload;
+import org.mule.common.metadata.DefaultMetaData;
+import org.mule.common.metadata.DefaultMetaDataKey;
+import org.mule.common.metadata.DefaultPojoMetaDataModel;
+import org.mule.common.metadata.MetaData;
+import org.mule.common.metadata.MetaDataKey;
 import org.mule.common.query.Query;
+import org.mule.module.magento.api.AxisPortProvider;
+import org.mule.module.magento.api.DefaultAxisPortProvider;
+import org.mule.module.magento.api.MagentoClientAdaptor;
+import org.mule.module.magento.api.MagentoException;
+import org.mule.module.magento.api.catalog.AxisMagentoCatalogClient;
+import org.mule.module.magento.api.catalog.MagentoCatalogClient;
 import org.mule.module.magento.api.catalog.model.MediaMimeType;
 import org.mule.module.magento.api.catalog.model.ProductIdentifiers;
+import org.mule.module.magento.api.customer.AxisMagentoInventoryClient;
+import org.mule.module.magento.api.customer.MagentoInventoryClient;
+import org.mule.module.magento.api.directory.AxisMagentoDirectoryClient;
+import org.mule.module.magento.api.directory.MagentoDirectoryClient;
+import org.mule.module.magento.api.inventory.AxisMagentoCustomerClient;
+import org.mule.module.magento.api.inventory.MagentoCustomerClient;
+import org.mule.module.magento.api.order.AxisMagentoOrderClient;
+import org.mule.module.magento.api.order.MagentoOrderClient;
 import org.mule.module.magento.api.order.model.Carrier;
-import org.mule.module.magento.connection.strategy.MagentoConnectionManagement;
-import org.mule.module.magento.metadata.category.MagentoMetadataCategory;
+import org.mule.module.magento.api.shoppingCart.AxisMagentoShoppingCartClient;
+import org.mule.module.magento.api.shoppingCart.MagentoShoppingCartClient;
 import org.mule.module.magento.query.MagentoQueryVisitor;
 
-import java.io.*;
-import java.util.List;
+import com.magento.api.*;
+
+//import java.util.Map;
 
 /**
  * Magento is a feature-rich eCommerce platform built on open-source technology that provides online merchants with
@@ -31,18 +72,173 @@ import java.util.List;
  *
  * @author MuleSoft, Inc.
  */
-@Connector(name = "magento", schemaVersion = "1.1", friendlyName = "Magento", minMuleVersion = "3.5")
+@Connector(name = "magento", schemaVersion = "1.1", friendlyName = "Magento", minMuleVersion = "3.5", metaData=MetaDataSwitch.ON)
 public class MagentoCloudConnector {
 
-    @ConnectionStrategy
-    private MagentoConnectionManagement connection;
+    private static final String ENTITY_PACKAGE = "com.magento.api";
 
-    public MagentoConnectionManagement getConnection() {
-        return connection;
+    private MagentoOrderClient<MagentoException> orderClient;
+    private MagentoCustomerClient<MagentoException> customerClient;
+    private MagentoInventoryClient<MagentoException> inventoryClient;
+    private MagentoDirectoryClient<MagentoException> directoryClient;
+    private MagentoCatalogClient<MagentoException> catalogClient;
+    private MagentoShoppingCartClient<MagentoException> shoppingCartClient;
+
+    /**
+     *
+     */
+    private String user;
+
+    /**
+     * The address to access Magento Web Services
+     */
+    private String serverAddress;
+
+    /**
+     *
+     * @param username The user name to access Magento Web Services
+     * @param password The password to access Magento Web Services
+     * @param address The address to access Magento Web Services
+     */
+    @Connect
+    public void initialiseConnector(@ConnectionKey String username, @Password String password, String address)
+            throws ConnectionException {
+        setUser(username);
+        setServerAddress(address);
+        PortProviderInitializer initializer = new PortProviderInitializer(username, password, address);
+        setOrderClient(new AxisMagentoOrderClient(initializer.getPortProvider()));
+        setCustomerClient(new AxisMagentoCustomerClient(initializer.getPortProvider()));
+        setInventoryClient(new AxisMagentoInventoryClient(initializer.getPortProvider()));
+        setDirectoryClient(new AxisMagentoDirectoryClient(initializer.getPortProvider()));
+        setCatalogClient(new AxisMagentoCatalogClient(initializer.getPortProvider()));
+        setShoppingCartClient(new AxisMagentoShoppingCartClient(initializer.getPortProvider()));
+
+        //Dummy operation for connectivity testing
+        try {
+            getCatalogCurrentStoreView();
+        } catch (Exception e) {
+            final ConnectionException connectionException = new ConnectionException(ConnectionExceptionCode.INCORRECT_CREDENTIALS, null, e.getMessage());
+            connectionException.initCause(e);
+            throw connectionException;
+
+        }
     }
 
-    public void setConnection(MagentoConnectionManagement connection) {
-        this.connection = connection;
+    @ValidateConnection
+    public boolean validateConnection() {
+        return orderClient != null && customerClient != null && inventoryClient != null && directoryClient != null &&
+                catalogClient != null && shoppingCartClient != null;
+    }
+
+    @Disconnect
+    public void disconnect() {
+        setOrderClient(null);
+        setCustomerClient(null);
+        setInventoryClient(null);
+        setDirectoryClient(null);
+        setCatalogClient(null);
+        setShoppingCartClient(null);
+    }
+
+    @ConnectionIdentifier
+    public String connectionId() {
+        return String.format("%s-at-%s", getUser(), getServerAddress());
+    }
+
+    private MetaDataKey createKey(Class<?> cls) {
+        return new DefaultMetaDataKey(cls.getSimpleName(), cls.getSimpleName());
+    }
+
+    @MetaDataKeyRetriever
+    public List<MetaDataKey> getMetadataKeys() {
+        List<MetaDataKey> keys = new ArrayList<MetaDataKey>();
+        keys.add(this.createKey(ApiEntity.class));
+        keys.add(this.createKey(ApiMethodEntity.class));
+        keys.add(this.createKey(AssociativeEntity.class));
+        keys.add(this.createKey(AssociativeMultiEntity.class));
+        keys.add(this.createKey(CatalogAttributeEntity.class));
+        keys.add(this.createKey(CatalogAttributeOptionEntity.class));
+        keys.add(this.createKey(CatalogCategoryEntity.class));
+        keys.add(this.createKey(CatalogInventoryStockItemEntity.class));
+        keys.add(this.createKey(CatalogInventoryStockItemUpdateEntity.class));
+        keys.add(this.createKey(CatalogProductAdditionalAttributesEntity.class));
+        keys.add(this.createKey(CatalogProductAttributeEntity.class));
+        keys.add(this.createKey(CatalogProductAttributeFrontendLabelEntity.class));
+        keys.add(this.createKey(CatalogProductAttributeMediaCreateEntity.class));
+        keys.add(this.createKey(CatalogProductAttributeMediaTypeEntity.class));
+        keys.add(this.createKey(CatalogProductAttributeOptionLabelEntity.class));
+        keys.add(this.createKey(CatalogProductAttributeSetEntity.class));
+        keys.add(this.createKey(CatalogProductCreateEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionAdditionalFieldsEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionInfoEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionListEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionTypesEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionValueAddEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionValueInfoEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionValueListEntity.class));
+        keys.add(this.createKey(CatalogProductCustomOptionValueUpdateEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkAddEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkAddSampleEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkFileEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkFileInfoEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkInfoEntity.class));
+        keys.add(this.createKey(CatalogProductDownloadableLinkSampleEntity.class));
+        keys.add(this.createKey(CatalogProductEntity.class));
+        keys.add(this.createKey(CatalogProductImageEntity.class));
+        keys.add(this.createKey(CatalogProductImageFileEntity.class));
+        keys.add(this.createKey(CatalogProductLinkAttributeEntity.class));
+        keys.add(this.createKey(CatalogProductLinkEntity.class));
+        keys.add(this.createKey(CatalogProductReturnEntity.class));
+        keys.add(this.createKey(CatalogProductTagAddEntity.class));
+        keys.add(this.createKey(CatalogProductTagInfoEntity.class));
+        keys.add(this.createKey(CatalogProductTagListEntity.class));
+        keys.add(this.createKey(CatalogProductTagUpdateEntity.class));
+        keys.add(this.createKey(CatalogProductTierPriceEntity.class));
+        keys.add(this.createKey(CatalogProductTypeEntity.class));
+        keys.add(this.createKey(CustomerCustomerEntity.class));
+        keys.add(this.createKey(CustomerGroupEntity.class));
+        keys.add(this.createKey(DirectoryCountryEntity.class));
+        keys.add(this.createKey(DirectoryRegionEntity.class));
+        keys.add(this.createKey(ExistsFaltureEntity.class));
+        keys.add(this.createKey(GiftMessageAssociativeProductsEntity.class));
+        keys.add(this.createKey(MagentoInfoEntity.class));
+        keys.add(this.createKey(SalesOrderAddressEntity.class));
+        keys.add(this.createKey(SalesOrderCreditmemoCommentEntity.class));
+        keys.add(this.createKey(SalesOrderCreditmemoEntity.class));
+        keys.add(this.createKey(SalesOrderCreditmemoItemEntity.class));
+        keys.add(this.createKey(SalesOrderEntity.class));
+        keys.add(this.createKey(SalesOrderInvoiceCommentEntity.class));
+        keys.add(this.createKey(SalesOrderInvoiceEntity.class));
+        keys.add(this.createKey(SalesOrderInvoiceItemEntity.class));
+        keys.add(this.createKey(SalesOrderItemEntity.class));
+        keys.add(this.createKey(SalesOrderListEntity.class));
+        keys.add(this.createKey(SalesOrderPaymentEntity.class));
+        keys.add(this.createKey(SalesOrderShipmentCommentEntity.class));
+        keys.add(this.createKey(SalesOrderShipmentEntity.class));
+        keys.add(this.createKey(SalesOrderShipmentItemEntity.class));
+        keys.add(this.createKey(SalesOrderShipmentTrackEntity.class));
+        keys.add(this.createKey(SalesOrderStatusHistoryEntity.class));
+        keys.add(this.createKey(ShoppingCartAddressEntity.class));
+        keys.add(this.createKey(ShoppingCartCustomerAddressEntity.class));
+        keys.add(this.createKey(ShoppingCartCustomerEntity.class));
+        keys.add(this.createKey(ShoppingCartInfoEntity.class));
+        keys.add(this.createKey(ShoppingCartItemEntity.class));
+        keys.add(this.createKey(ShoppingCartLicenseEntity.class));
+        keys.add(this.createKey(ShoppingCartPaymentEntity.class));
+        keys.add(this.createKey(ShoppingCartPaymentMethodEntity.class));
+        keys.add(this.createKey(ShoppingCartProductEntity.class));
+        keys.add(this.createKey(ShoppingCartShippingMethodEntity.class));
+        keys.add(this.createKey(ShoppingCartTotalsEntity.class));
+        keys.add(this.createKey(StoreEntity.class));
+
+        return keys;
+    }
+
+    @MetaDataRetriever
+    public MetaData getMetadata(MetaDataKey key) throws ClassNotFoundException {
+        Class<?> clazz = Class.forName(String.format("%s.%s", ENTITY_PACKAGE, key.getId()));
+        return new DefaultMetaData(new DefaultPojoMetaDataModel(clazz));
     }
 
     /**
@@ -57,14 +253,13 @@ public class MagentoCloudConnector {
      * @return true if it was successful
      */
     @Processor
-    @MetaDataScope(MagentoMetadataCategory.class)
     public boolean addOrderShipmentComment(String shipmentId,
                                            String comment,
                                            @Default("false") boolean sendEmail,
                                            @Default("false") boolean includeCommentInEmail)
 
     {
-        return connection.getOrderClient().addOrderShipmentComment(shipmentId, comment, sendEmail, includeCommentInEmail);
+        return orderClient.addOrderShipmentComment(shipmentId, comment, sendEmail, includeCommentInEmail);
     }
 
     /**
@@ -83,7 +278,7 @@ public class MagentoCloudConnector {
                                      String carrierCode,
                                      String title,
                                      String trackId) {
-        return connection.getOrderClient().addOrderShipmentTrack(shipmentId, carrierCode, title, trackId);
+        return orderClient.addOrderShipmentTrack(shipmentId, carrierCode, title, trackId);
     }
 
     /**
@@ -96,7 +291,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean cancelOrder(String orderId) {
-        return connection.getOrderClient().cancelOrder(orderId);
+        return orderClient.cancelOrder(orderId);
     }
 
     /**
@@ -117,7 +312,7 @@ public class MagentoCloudConnector {
                                       @Optional String comment,
                                       @Default("false") boolean sendEmail,
                                       @Default("false") boolean includeCommentInEmail) {
-        return connection.getOrderClient().createOrderShipment(orderId, itemsQuantities, comment, sendEmail, includeCommentInEmail);
+        return orderClient.createOrderShipment(orderId, itemsQuantities, comment, sendEmail, includeCommentInEmail);
     }
 
     /**
@@ -130,7 +325,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public SalesOrderEntity getOrder(String orderId) {
-        return connection.getOrderClient().getOrder(orderId);
+        return orderClient.getOrder(orderId);
     }
 
     /**
@@ -138,12 +333,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getOrderInvoice}
      *
+     *
      * @param invoiceId the target invoiceId
      * @return the invoice attributes
      */
     @Processor
     public SalesOrderInvoiceEntity getOrderInvoice(String invoiceId) {
-        return connection.getOrderClient().getOrderInvoice(invoiceId);
+        return orderClient.getOrderInvoice(invoiceId);
     }
 
     /**
@@ -156,7 +352,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<Carrier> getOrderShipmentCarriers(String orderId) {
-        return connection.getOrderClient().getOrderShipmentCarriers(orderId);
+        return orderClient.getOrderShipmentCarriers(orderId);
     }
 
     /**
@@ -164,12 +360,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getOrderShipment}
      *
+     *
      * @param shipmentId the id of the shipment
      * @return the shipment attributes
      */
     @Processor
     public SalesOrderShipmentEntity getOrderShipment(String shipmentId) {
-        return connection.getOrderClient().getOrderShipment(shipmentId);
+        return orderClient.getOrderShipment(shipmentId);
     }
 
     /**
@@ -182,13 +379,15 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean holdOrder(String orderId) {
-        return connection.getOrderClient().holdOrder(orderId);
+        return orderClient.holdOrder(orderId);
     }
 
     /**
      * Lists order attributes that match the given filtering expression.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listOrders}
+     *
+     *
      *
      * @param filter optional filtering expression - one or more comma-separated
      *               unary or binary predicates, one for each filter, in the form
@@ -200,7 +399,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<SalesOrderListEntity> listOrders(@org.mule.api.annotations.Query @Optional String filter) {
-        return connection.getOrderClient().listOrders(filter);
+        return orderClient.listOrders(filter);
     }
 
     @QueryTranslator
@@ -216,6 +415,7 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listOrdersInvoices}
      *
+     *
      * @param filter optional filtering expression - one or more comma-separated
      *               unary or binary predicates, one for each filter, in the form
      *               filterType(attributeName, value), for binary filters or
@@ -226,7 +426,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<SalesOrderInvoiceEntity> listOrdersInvoices(@org.mule.api.annotations.Query @Optional String filter) {
-        return connection.getOrderClient().listOrdersInvoices(filter);
+        return orderClient.listOrdersInvoices(filter);
     }
 
     /**
@@ -234,6 +434,7 @@ public class MagentoCloudConnector {
      * optional filtering expression
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listOrdersShipments}
+     *
      *
      * @param filter optional filtering expression - one or more comma-separated
      *               unary or binary predicates, one for each filter, in the form
@@ -245,7 +446,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<SalesOrderShipmentEntity> listOrdersShipments(@org.mule.api.annotations.Query @Optional String filter) {
-        return connection.getOrderClient().listOrdersShipments(filter);
+        return orderClient.listOrdersShipments(filter);
     }
 
     /**
@@ -259,7 +460,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean deleteOrderShipmentTrack(String shipmentId, String trackId) {
-        return connection.getOrderClient().deleteOrderShipmentTrack(shipmentId, trackId);
+        return orderClient.deleteOrderShipmentTrack(shipmentId, trackId);
     }
 
     /**
@@ -278,7 +479,7 @@ public class MagentoCloudConnector {
                                    String status,
                                    String comment,
                                    @Default("false") boolean sendEmail) {
-        return connection.getOrderClient().addOrderComment(orderId, status, comment, sendEmail);
+        return orderClient.addOrderComment(orderId, status, comment, sendEmail);
     }
 
     /**
@@ -288,10 +489,11 @@ public class MagentoCloudConnector {
      *
      * @param orderId the id of the order to remove from hold state
      * @return true if the operation was successful
+     *
      */
     @Processor
     public boolean unholdOrder(String orderId) {
-        return connection.getOrderClient().unholdOrder(orderId);
+        return orderClient.unholdOrder(orderId);
     }
 
     /**
@@ -312,7 +514,7 @@ public class MagentoCloudConnector {
                                      @Optional String comment,
                                      @Default("false") boolean sendEmail,
                                      @Default("false") boolean includeCommentInEmail) {
-        return connection.getOrderClient().createOrderInvoice(orderId, itemsQuantities, comment, sendEmail,
+        return orderClient.createOrderInvoice(orderId, itemsQuantities, comment, sendEmail,
                 includeCommentInEmail);
     }
 
@@ -332,7 +534,7 @@ public class MagentoCloudConnector {
                                           String comment,
                                           @Default("false") boolean sendEmail,
                                           @Default("false") boolean includeCommentInEmail) {
-        return connection.getOrderClient().addOrderInvoiceComment(invoiceId, comment, sendEmail, includeCommentInEmail);
+        return orderClient.addOrderInvoiceComment(invoiceId, comment, sendEmail, includeCommentInEmail);
     }
 
     /**
@@ -344,7 +546,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public void captureOrderInvoice(String invoiceId) {
-        connection.getOrderClient().captureOrderInvoice(invoiceId);
+        orderClient.captureOrderInvoice(invoiceId);
     }
 
     /**
@@ -357,7 +559,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean voidOrderInvoice(String invoiceId) {
-        return connection.getOrderClient().voidOrderInvoice(invoiceId);
+        return orderClient.voidOrderInvoice(invoiceId);
     }
 
     /**
@@ -369,7 +571,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public void cancelOrderInvoice(String invoiceId) {
-        connection.getOrderClient().cancelOrderInvoice(invoiceId);
+        orderClient.cancelOrderInvoice(invoiceId);
     }
 
     /**
@@ -378,14 +580,14 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:createCustomerAddress}
      *
-     * @param customerId      the customer
+     * @param customerId the customer
      * @param customerAddress the address attributes
      * @return a new customer address id
      */
     @Processor
     public int createCustomerAddress(int customerId, @Default("#[payload]")
     @Placement(group = "Address Attributes") CustomerAddressEntityCreate customerAddress) {
-        return connection.getCustomerClient().createCustomerAddress(customerId, customerAddress);
+        return customerClient.createCustomerAddress(customerId, customerAddress);
     }
 
     /**
@@ -398,7 +600,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public int createCustomer(@Default("#[payload]") @Placement(group = "Customer Attributes") CustomerCustomerEntityToCreate customer) {
-        return connection.getCustomerClient().createCustomer(customer);
+        return customerClient.createCustomer(customer);
     }
 
     /**
@@ -411,7 +613,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean deleteCustomer(int customerId) {
-        return connection.getCustomerClient().deleteCustomer(customerId);
+        return customerClient.deleteCustomer(customerId);
     }
 
     /**
@@ -424,7 +626,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean deleteCustomerAddress(int addressId) {
-        return connection.getCustomerClient().deleteCustomerAddress(addressId);
+        return customerClient.deleteCustomerAddress(addressId);
     }
 
     /**
@@ -439,7 +641,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public CustomerCustomerEntity getCustomer(int customerId, @Placement(group = "Attributes to Retrieve") List<String> attributeNames) {
-        return connection.getCustomerClient().getCustomer(customerId, attributeNames);
+        return customerClient.getCustomer(customerId, attributeNames);
     }
 
     /**
@@ -447,12 +649,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getCustomerAddress}
      *
+     *
      * @param addressId the id of the address
      * @return the customer address attributes map
      */
     @Processor
     public CustomerAddressEntityItem getCustomerAddress(int addressId) {
-        return connection.getCustomerClient().getCustomerAddress(addressId);
+        return customerClient.getCustomerAddress(addressId);
     }
 
     /**
@@ -460,12 +663,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listCustomerAddresses}
      *
+     *
      * @param customerId the id of the customer
      * @return a listing of addresses
      */
     @Processor
     public List<CustomerAddressEntityItem> listCustomerAddresses(int customerId) {
-        return connection.getCustomerClient().listCustomerAddresses(customerId);
+        return customerClient.listCustomerAddresses(customerId);
     }
 
     /**
@@ -477,13 +681,14 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<CustomerGroupEntity> listCustomerGroups() {
-        return connection.getCustomerClient().listCustomerGroups();
+        return customerClient.listCustomerGroups();
     }
 
     /**
      * Answers a list of customer attributes for the given filter expression.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listCustomers}
+     *
      *
      * @param filter optional filtering expression - one or more comma-separated
      *               unary or binary predicates, one for each filter, in the form
@@ -495,7 +700,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<CustomerCustomerEntity> listCustomers(@org.mule.api.annotations.Query @Optional String filter) {
-        return connection.getCustomerClient().listCustomers(filter);
+        return customerClient.listCustomers(filter);
     }
 
     /**
@@ -505,13 +710,13 @@ public class MagentoCloudConnector {
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateCustomer}
      *
      * @param customerId the target customer to update
-     * @param customer   the customer attributes
+     * @param customer the customer attributes
      * @return update result
      */
     @Processor
     public boolean updateCustomer(int customerId,
                                   @Default("#[payload]") @Placement(group = "Customer Attributes to Update") CustomerCustomerEntityToCreate customer) {
-        return connection.getCustomerClient().updateCustomer(customerId, customer);
+        return customerClient.updateCustomer(customerId, customer);
     }
 
     /**
@@ -519,13 +724,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateCustomerAddress}
      *
-     * @param addressId       the customer address to update
+     * @param addressId  the customer address to update
      * @param customerAddress the address attributes to update
      * @return update result
      */
     @Processor
     public boolean updateCustomerAddress(int addressId, @Default("#[payload]") @Placement(group = "Address Attributes to Update") CustomerAddressEntityCreate customerAddress) {
-        return connection.getCustomerClient().updateCustomerAddress(addressId, customerAddress);
+        return customerClient.updateCustomerAddress(addressId, customerAddress);
     }
 
     /**
@@ -533,12 +738,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listStockItems}
      *
+     *
      * @param productIdOrSkus a not empty list of product ids or skus whose attributes to list
      * @return a list of stock items attributes
      */
     @Processor
     public List<CatalogInventoryStockItemEntity> listStockItems(@Placement(group = "Product Ids or SKUs") List<String> productIdOrSkus) {
-        return connection.getInventoryClient().listStockItems(productIdOrSkus);
+        return inventoryClient.listStockItems(productIdOrSkus);
     }
 
 
@@ -548,13 +754,13 @@ public class MagentoCloudConnector {
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateStockItem}
      *
      * @param productIdOrSku the product id or sku of the product to update
-     * @param stockItem      the attributes to update
+     * @param stockItem     the attributes to update
      * @return update result
      */
     @Processor
     public int updateStockItem(String productIdOrSku,
                                @Default("#[payload]") @Placement(group = "Attributes to Update") CatalogInventoryStockItemUpdateEntity stockItem) {
-        return connection.getInventoryClient().updateStockItem(productIdOrSku, stockItem);
+        return inventoryClient.updateStockItem(productIdOrSku, stockItem);
     }
 
     /**
@@ -566,7 +772,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<DirectoryCountryEntity> listDirectoryCountries() {
-        return connection.getDirectoryClient().listDirectoryCountries();
+        return directoryClient.listDirectoryCountries();
     }
 
     /**
@@ -574,12 +780,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listDirectoryRegions}
      *
+     *
      * @param countryId the country code, in ISO2 or ISO3 format
      * @return the collection of regions attributes
      */
     @Processor
     public List<DirectoryRegionEntity> listDirectoryRegions(String countryId) {
-        return connection.getDirectoryClient().listDirectoryRegions(countryId);
+        return directoryClient.listDirectoryRegions(countryId);
     }
 
 
@@ -597,7 +804,7 @@ public class MagentoCloudConnector {
      *                             product sku
      * @param productIdOrSku       the id or sku of the source product.
      * @param linkedProductIdOrSku the destination product id or sku.
-     * @param productLinkEntity    the link attributes
+     * @param productLinkEntity           the link attributes
      * @return true if the operation was successful
      */
     @Processor
@@ -607,7 +814,7 @@ public class MagentoCloudConnector {
                                   @Optional String productIdOrSku,
                                   String linkedProductIdOrSku,
                                   @Default("#[payload]") @Placement(group = "Address Attributes to Update") CatalogProductLinkEntity productLinkEntity) {
-        return connection.getCatalogClient().addProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku), linkedProductIdOrSku,
+        return catalogClient.addProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku), linkedProductIdOrSku,
                 productLinkEntity);
     }
 
@@ -617,18 +824,18 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:createProductAttributeMedia}
      *
-     * @param productId                          the id of the product. Use it instead of productIdOrSku
-     *                                           in case you are sure the product identifier is a
-     *                                           product id
-     * @param productSku                         the sku of the product. Use it instead of productIdOrSku
-     *                                           in case you are sure the product identifier is a
-     *                                           product sku
-     * @param productIdOrSku                     the id or sku of the product.
-     * @param catalogProductAttributeMediaEntity the media attributes
-     * @param storeViewIdOrCode                  the id or code of the target store. Left unspecified for using current store
-     * @param payload                            the image to upload. It may be a file, an input stream or a byte array
-     * @param mimeType                           the mimetype
-     * @param baseFileName                       the base name of the new remote image
+     * @param productId         the id of the product. Use it instead of productIdOrSku
+     *                          in case you are sure the product identifier is a
+     *                          product id
+     * @param productSku        the sku of the product. Use it instead of productIdOrSku
+     *                          in case you are sure the product identifier is a
+     *                          product sku
+     * @param productIdOrSku    the id or sku of the product.
+     * @param catalogProductAttributeMediaEntity        the media attributes
+     * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
+     * @param payload           the image to upload. It may be a file, an input stream or a byte array
+     * @param mimeType          the mimetype
+     * @param baseFileName      the base name of the new remote image
      * @return the new image filename
      * @throws FileNotFoundException when the payload object is not a valid file
      */
@@ -636,12 +843,12 @@ public class MagentoCloudConnector {
     public String createProductAttributeMedia(@Optional Integer productId,
                                               @Optional String productSku,
                                               @Optional String productIdOrSku,
-                                              @Placement(group = "Media Attributes") CatalogProductAttributeMediaCreateEntity catalogProductAttributeMediaEntity,
+                                              @Default("#[payload]") @Placement(group = "Media Attributes") CatalogProductAttributeMediaCreateEntity catalogProductAttributeMediaEntity,
                                               @Optional String storeViewIdOrCode,
                                               @Payload Object payload,
                                               MediaMimeType mimeType,
                                               @Optional String baseFileName) throws FileNotFoundException {
-        return connection.getCatalogClient().createProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku),
+        return catalogClient.createProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku),
                 catalogProductAttributeMediaEntity, createContent(payload), mimeType, baseFileName, storeViewIdOrCode);
     }
 
@@ -681,7 +888,7 @@ public class MagentoCloudConnector {
                                                String fileName)
 
     {
-        return connection.getCatalogClient().deleteProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), fileName);
+        return catalogClient.deleteProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), fileName);
     }
 
     /**
@@ -706,13 +913,14 @@ public class MagentoCloudConnector {
                                      @Optional String productSku,
                                      @Optional String productIdOrSku,
                                      String linkedProductIdOrSku) {
-        return connection.getCatalogClient().deleteProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku), linkedProductIdOrSku);
+        return catalogClient.deleteProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku), linkedProductIdOrSku);
     }
 
     /**
      * Lists linked products to the given product and for the given link type.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getProductAttributeMedia}
+     *
      *
      * @param productId         the id of the product. Use it instead of productIdOrSku in
      *                          case you are sure the product identifier is a product id
@@ -729,7 +937,7 @@ public class MagentoCloudConnector {
                                                               @Optional String productIdOrSku,
                                                               String fileName,
                                                               @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().getProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), fileName, storeViewIdOrCode);
+        return catalogClient.getProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), fileName, storeViewIdOrCode);
     }
 
 
@@ -742,7 +950,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public int getCatalogCurrentStoreView() {
-        return connection.getCatalogClient().getCatalogCurrentStoreView();
+        return catalogClient.getCatalogCurrentStoreView();
     }
 
     /**
@@ -757,7 +965,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public int updateCategoryAttributeStoreView(String storeViewIdOrCode) {
-        return connection.getCatalogClient().updateCatalogCurrentStoreView(storeViewIdOrCode);
+        return catalogClient.updateCatalogCurrentStoreView(storeViewIdOrCode);
     }
 
     /**
@@ -770,7 +978,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<CatalogAttributeEntity> listCategoryAttributes() {
-        return connection.getCatalogClient().listCategoryAttributes();
+        return catalogClient.listCategoryAttributes();
     }
 
     /**
@@ -779,6 +987,7 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listCategoryAttributeOptions}
      *
+     *
      * @param attributeId       the target attribute whose options will be retrieved
      * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
      * @return the list of category attribute options
@@ -786,7 +995,7 @@ public class MagentoCloudConnector {
     @Processor
     public List<CatalogAttributeOptionEntity> listCategoryAttributeOptions(String attributeId,
                                                                            @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().listCategoryAttributeOptions(attributeId, storeViewIdOrCode);
+        return catalogClient.listCategoryAttributeOptions(attributeId, storeViewIdOrCode);
     }
 
     /**
@@ -794,6 +1003,7 @@ public class MagentoCloudConnector {
      * method
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductAttributeMedia}
+     *
      *
      * @param productId         the id of the product. Use it instead of productIdOrSku in
      *                          case you are sure the product identifier is a product id
@@ -808,7 +1018,7 @@ public class MagentoCloudConnector {
                                                                      @Optional String productSku,
                                                                      @Optional String productIdOrSku,
                                                                      @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().listProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), storeViewIdOrCode);
+        return catalogClient.listProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), storeViewIdOrCode);
     }
 
     /**
@@ -817,12 +1027,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductAttributeMediaTypes}
      *
+     *
      * @param setId the setId
      * @return the list of attribute media types
      */
     @Processor
     public List<CatalogProductAttributeMediaTypeEntity> listProductAttributeMediaTypes(int setId) {
-        return connection.getCatalogClient().listProductAttributeMediaTypes(setId);
+        return catalogClient.listProductAttributeMediaTypes(setId);
     }
 
     /**
@@ -831,6 +1042,7 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductAttributeOptions}
      *
+     *
      * @param attributeId       the target attribute whose options will be listed
      * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
      * @return the attributes list
@@ -838,7 +1050,7 @@ public class MagentoCloudConnector {
     @Processor
     public List<CatalogAttributeOptionEntity> listProductAttributeOptions(String attributeId,
                                                                           @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().listProductAttributeOptions(attributeId, storeViewIdOrCode);
+        return catalogClient.listProductAttributeOptions(attributeId, storeViewIdOrCode);
     }
 
 
@@ -848,12 +1060,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductAttributes}
      *
+     *
      * @param setId the set id
      * @return the list of product attributes
      */
     @Processor
     public List<CatalogAttributeEntity> listProductAttributes(int setId) {
-        return connection.getCatalogClient().listProductAttributes(setId);
+        return catalogClient.listProductAttributes(setId);
     }
 
     /**
@@ -866,7 +1079,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<CatalogProductAttributeSetEntity> listProductAttributeSets() {
-        return connection.getCatalogClient().listProductAttributeSets();
+        return catalogClient.listProductAttributeSets();
     }
 
     /**
@@ -874,6 +1087,7 @@ public class MagentoCloudConnector {
      * SOAP Method.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductAttributeTierPrices}
+     *
      *
      * @param productId      the id of the product. Use it instead of productIdOrSku in
      *                       case you are sure the product identifier is a product id
@@ -886,13 +1100,14 @@ public class MagentoCloudConnector {
     public List<CatalogProductTierPriceEntity> listProductAttributeTierPrices(@Optional Integer productId,
                                                                               @Optional String productSku,
                                                                               @Optional String productIdOrSku) {
-        return connection.getCatalogClient().listProductAttributeTierPrices(ProductIdentifiers.from(productSku, productId, productIdOrSku));
+        return catalogClient.listProductAttributeTierPrices(ProductIdentifiers.from(productSku, productId, productIdOrSku));
     }
 
     /**
      * Lists linked products to the given product and for the given link type
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductLink}
+     *
      *
      * @param type           the link type
      * @param productId      the id of the product. Use it instead of productIdOrSku in
@@ -907,7 +1122,7 @@ public class MagentoCloudConnector {
                                                           @Optional Integer productId,
                                                           @Optional String productSku,
                                                           @Optional String productIdOrSku) {
-        return connection.getCatalogClient().listProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku));
+        return catalogClient.listProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku));
     }
 
 
@@ -916,12 +1131,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProductLinkAttributes}
      *
+     *
      * @param type the product type
      * @return the listing of product attributes
      */
     @Processor
     public List<CatalogProductLinkAttributeEntity> listProductLinkAttributes(String type) {
-        return connection.getCatalogClient().listProductLinkAttributes(type);
+        return catalogClient.listProductLinkAttributes(type);
     }
 
     /**
@@ -933,7 +1149,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<String> listProductLinkTypes() {
-        return connection.getCatalogClient().listProductLinkTypes();
+        return catalogClient.listProductLinkTypes();
     }
 
     /**
@@ -945,7 +1161,7 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<CatalogProductTypeEntity> listProductTypes() {
-        return connection.getCatalogClient().listProductTypes();
+        return catalogClient.listProductTypes();
     }
 
     /**
@@ -953,14 +1169,14 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateProductAttributeMedia}
      *
-     * @param productId                          the id of the product. Use it instead of productIdOrSku in
-     *                                           case you are sure the product identifier is a product id
-     * @param productSku                         the sku of the product. Use it instead of productIdOrSku in
-     *                                           case you are sure the product identifier is a product sku
-     * @param productIdOrSku                     the id or sku of the product.
-     * @param fileName                           the name of the remote media file to update
-     * @param catalogProductAttributeMediaEntity the attributes to update
-     * @param storeViewIdOrCode                  the id or code of the target store. Left unspecified for using current store
+     * @param productId         the id of the product. Use it instead of productIdOrSku in
+     *                          case you are sure the product identifier is a product id
+     * @param productSku        the sku of the product. Use it instead of productIdOrSku in
+     *                          case you are sure the product identifier is a product sku
+     * @param productIdOrSku    the id or sku of the product.
+     * @param fileName          the name of the remote media file to update
+     * @param catalogProductAttributeMediaEntity        the attributes to update
+     * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
      * @return update result
      */
     @Processor
@@ -970,7 +1186,7 @@ public class MagentoCloudConnector {
                                                String fileName,
                                                @Default("#[payload]") @Placement(group = "Media Attributes to Update") CatalogProductAttributeMediaCreateEntity catalogProductAttributeMediaEntity,
                                                @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().updateProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), fileName,
+        return catalogClient.updateProductAttributeMedia(ProductIdentifiers.from(productSku, productId, productIdOrSku), fileName,
                 catalogProductAttributeMediaEntity, storeViewIdOrCode);
     }
 
@@ -981,12 +1197,12 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateProductAttributeTierPrice}
      *
-     * @param productId                     the id of the product. Use it instead of productIdOrSku in
-     *                                      case you are sure the product identifier is a product id
-     * @param productSku                    the sku of the product. Use it instead of productIdOrSku in
-     *                                      case you are sure the product identifier is a product sku
-     * @param productIdOrSku                the id or sku of the product.
-     * @param catalogProductTierPriceEntity the tier price to update.
+     * @param productId      the id of the product. Use it instead of productIdOrSku in
+     *                       case you are sure the product identifier is a product id
+     * @param productSku     the sku of the product. Use it instead of productIdOrSku in
+     *                       case you are sure the product identifier is a product sku
+     * @param productIdOrSku the id or sku of the product.
+     * @param catalogProductTierPriceEntity     the tier price to update.
      * @return update result
      */
     @Processor
@@ -994,7 +1210,7 @@ public class MagentoCloudConnector {
                                                @Optional String productSku,
                                                @Optional String productIdOrSku,
                                                @Default("#[payload]") @Placement(group = "Tier Price Attributes to Update") CatalogProductTierPriceEntity catalogProductTierPriceEntity) {
-        return connection.getCatalogClient().updateProductAttributeTierPrice(ProductIdentifiers.from(productSku, productId, productIdOrSku), catalogProductTierPriceEntity);
+        return catalogClient.updateProductAttributeTierPrice(ProductIdentifiers.from(productSku, productId, productIdOrSku), catalogProductTierPriceEntity);
     }
 
 
@@ -1003,16 +1219,16 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateProductLink}
      *
-     * @param type                     the link type
-     * @param productId                the id of the source product. Use it instead of productIdOrSku
-     *                                 in case you are sure the source product identifier is a
-     *                                 product id
-     * @param productSku               the sku of the source product. Use it instead of productIdOrSku
-     *                                 in case you are sure the source product identifier is a
-     *                                 product sku
-     * @param productIdOrSku           the id or sku of the source product.
-     * @param linkedProductIdOrSku     the destination product id or sku.
-     * @param catalogProductLinkEntity the link attributes
+     * @param type                 the link type
+     * @param productId            the id of the source product. Use it instead of productIdOrSku
+     *                             in case you are sure the source product identifier is a
+     *                             product id
+     * @param productSku           the sku of the source product. Use it instead of productIdOrSku
+     *                             in case you are sure the source product identifier is a
+     *                             product sku
+     * @param productIdOrSku       the id or sku of the source product.
+     * @param linkedProductIdOrSku the destination product id or sku.
+     * @param catalogProductLinkEntity           the link attributes
      * @return update result
      */
     @Processor
@@ -1022,7 +1238,7 @@ public class MagentoCloudConnector {
                                      @Optional String productIdOrSku,
                                      String linkedProductIdOrSku,
                                      @Default("#[payload]") @Placement(group = "Link Attributes to Update") CatalogProductLinkEntity catalogProductLinkEntity) {
-        return connection.getCatalogClient().updateProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku), linkedProductIdOrSku,
+        return catalogClient.updateProductLink(type, ProductIdentifiers.from(productSku, productId, productIdOrSku), linkedProductIdOrSku,
                 catalogProductLinkEntity);
     }
 
@@ -1031,12 +1247,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listCategoryProducts}
      *
+     *
      * @param categoryId the category
      * @return the listing of category products
      */
     @Processor
     public List<CatalogAssignedProduct> listCategoryProducts(int categoryId) {
-        return connection.getCatalogClient().listCategoryProducts(categoryId);
+        return catalogClient.listCategoryProducts(categoryId);
     }
 
     /**
@@ -1059,7 +1276,7 @@ public class MagentoCloudConnector {
                                       @Optional String productSku,
                                       @Optional String productIdOrSku,
                                       @Optional String position) {
-        return connection.getCatalogClient().addCategoryProduct(
+        return catalogClient.addCategoryProduct(
                 categoryId, ProductIdentifiers.from(productSku, productId, productIdOrSku), position);
     }
 
@@ -1068,16 +1285,16 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:createCategory}
      *
-     * @param parentId              the parent category id
-     * @param catalogCategoryEntity the new category attributes
-     * @param storeViewIdOrCode     the id or code of the target store. Left unspecified for using current store
+     * @param parentId          the parent category id
+     * @param catalogCategoryEntity        the new category attributes
+     * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
      * @return the new category id
      */
     @Processor
     public int createCategory(int parentId,
                               @Default("#[payload]") @Placement(group = "Category Attributes") CatalogCategoryEntityCreate catalogCategoryEntity,
                               @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().createCategory(parentId, catalogCategoryEntity, storeViewIdOrCode);
+        return catalogClient.createCategory(parentId, catalogCategoryEntity, storeViewIdOrCode);
     }
 
     /**
@@ -1090,13 +1307,14 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean deleteCategory(int categoryId) {
-        return connection.getCatalogClient().deleteCategory(categoryId);
+        return catalogClient.deleteCategory(categoryId);
     }
 
     /**
      * Answers category attributes. See catalog-category-info SOAP method.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getCategory}
+     *
      *
      * @param categoryId        the category whose attributes will be retrieved
      * @param storeViewIdOrCode the id or code of the target store. Left unspecified
@@ -1108,13 +1326,14 @@ public class MagentoCloudConnector {
     public CatalogCategoryInfo getCategory(int categoryId,
                                            @Optional String storeViewIdOrCode,
                                            @Placement(group = "Attribute Names to Retrieve") List<String> attributeNames) {
-        return connection.getCatalogClient().getCategory(categoryId, storeViewIdOrCode, attributeNames);
+        return catalogClient.getCategory(categoryId, storeViewIdOrCode, attributeNames);
     }
 
     /**
      * Answers levels of categories for a website, store view and parent category
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listCategoryLevels}
+     *
      *
      * @param website           the website
      * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
@@ -1127,7 +1346,7 @@ public class MagentoCloudConnector {
                                                                     @Optional String parentCategoryId)
 
     {
-        return connection.getCatalogClient().listCategoryLevels(website, storeViewIdOrCode, parentCategoryId);
+        return catalogClient.listCategoryLevels(website, storeViewIdOrCode, parentCategoryId);
     }
 
     /**
@@ -1146,7 +1365,7 @@ public class MagentoCloudConnector {
     public void moveCategory(int categoryId,
                              int parentId,
                              @Optional String afterId) {
-        connection.getCatalogClient().moveCategory(categoryId, parentId, afterId);
+        catalogClient.moveCategory(categoryId, parentId, afterId);
     }
 
     /**
@@ -1167,13 +1386,14 @@ public class MagentoCloudConnector {
                                          @Optional Integer productId,
                                          @Optional String productSku,
                                          @Optional String productIdOrSku) {
-        return connection.getCatalogClient().deleteCategoryProduct(categoryId, ProductIdentifiers.from(productSku, productId, productIdOrSku));
+        return catalogClient.deleteCategoryProduct(categoryId, ProductIdentifiers.from(productSku, productId, productIdOrSku));
     }
 
     /**
      * Answers the category tree. See catalog-category-tree SOAP method.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getCategoryTree}
+     *
      *
      * @param parentId          the parent id
      * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
@@ -1184,7 +1404,7 @@ public class MagentoCloudConnector {
                                                @Optional String storeViewIdOrCode)
 
     {
-        return connection.getCatalogClient().getCategoryTree(parentId, storeViewIdOrCode);
+        return catalogClient.getCategoryTree(parentId, storeViewIdOrCode);
     }
 
     /**
@@ -1192,16 +1412,16 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateCategory}
      *
-     * @param categoryId            the category to update
-     * @param catalogCategoryEntity the category new attributes
-     * @param storeViewIdOrCode     the id or code of the target store. Left unspecified for using current store
+     * @param categoryId        the category to update
+     * @param catalogCategoryEntity        the category new attributes
+     * @param storeViewIdOrCode the id or code of the target store. Left unspecified for using current store
      * @return update results
      */
     @Processor
     public boolean updateCategory(int categoryId,
                                   @Default("#[payload]") @Placement(group = "Category Attributes to Update") CatalogCategoryEntityCreate catalogCategoryEntity,
                                   @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().updateCategory(categoryId, catalogCategoryEntity, storeViewIdOrCode);
+        return catalogClient.updateCategory(categoryId, catalogCategoryEntity, storeViewIdOrCode);
     }
 
     /**
@@ -1224,7 +1444,7 @@ public class MagentoCloudConnector {
                                          @Optional String productSku,
                                          @Optional String productIdOrSku,
                                          String position) {
-        return connection.getCatalogClient().updateCategoryProduct(categoryId, ProductIdentifiers.from(productSku, productId, productIdOrSku), position);
+        return catalogClient.updateCategoryProduct(categoryId, ProductIdentifiers.from(productSku, productId, productIdOrSku), position);
     }
 
     /**
@@ -1232,12 +1452,13 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listInventoryStockItems}
      *
+     *
      * @param productIdOrSkus the list of product ids and/or skus whose attributes will be retrieved
      * @return the list of attributes
      */
     @Processor
     public List<CatalogInventoryStockItemEntity> listInventoryStockItems(@Placement(group = "Product Ids or SKUs") List<String> productIdOrSkus) {
-        return connection.getCatalogClient().listInventoryStockItems(productIdOrSkus);
+        return catalogClient.listInventoryStockItems(productIdOrSkus);
     }
 
     /**
@@ -1245,12 +1466,12 @@ public class MagentoCloudConnector {
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateInventoryStockItem}
      *
-     * @param productId                 the id of the product. Use it instead of productIdOrSku in
-     *                                  case you are sure the product identifier is a product id
-     * @param productSku                the sku of the product. Use it instead of productIdOrSku in
-     *                                  case you are sure the product identifier is a product sku
-     * @param productIdOrSku            the id or sku of the product.
-     * @param catalogInventoryStockItem the new attributes of the stock item
+     * @param productId      the id of the product. Use it instead of productIdOrSku in
+     *                       case you are sure the product identifier is a product id
+     * @param productSku     the sku of the product. Use it instead of productIdOrSku in
+     *                       case you are sure the product identifier is a product sku
+     * @param productIdOrSku the id or sku of the product.
+     * @param catalogInventoryStockItem     the new attributes of the stock item
      * @return update result
      */
     @Processor
@@ -1258,7 +1479,7 @@ public class MagentoCloudConnector {
                                         @Optional String productSku,
                                         @Optional String productIdOrSku,
                                         @Default("#[payload]") @Placement(group = "Stock Item Attributes") CatalogInventoryStockItemUpdateEntity catalogInventoryStockItem) {
-        return connection.getCatalogClient().updateInventoryStockItem(ProductIdentifiers.from(productSku, productId, productIdOrSku), catalogInventoryStockItem);
+        return catalogClient.updateInventoryStockItem(ProductIdentifiers.from(productSku, productId, productIdOrSku), catalogInventoryStockItem);
     }
 
     /**
@@ -1281,7 +1502,7 @@ public class MagentoCloudConnector {
                              @Default("#[payload]") @Placement(group = "Standard Product Attributes") CatalogProductCreateEntity attributes,
                              @Placement(group = "Non-standard Product Attributes") @Optional List<AssociativeEntity> additionalAttributes,
                              @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().createProduct(type, set, sku, attributes, additionalAttributes, storeViewIdOrCode);
+        return catalogClient.createProduct(type, set, sku, attributes, additionalAttributes, storeViewIdOrCode);
     }
 
 
@@ -1301,7 +1522,7 @@ public class MagentoCloudConnector {
     public int deleteProduct(@Optional Integer productId,
                              @Optional String productSku,
                              @Optional String productIdOrSku) {
-        return connection.getCatalogClient().deleteProduct(ProductIdentifiers.from(productSku, productId, productIdOrSku));
+        return catalogClient.deleteProduct(ProductIdentifiers.from(productSku, productId, productIdOrSku));
     }
 
 
@@ -1309,6 +1530,7 @@ public class MagentoCloudConnector {
      * Answers a product special price. See catalog-product-getSpecialPrice SOAP method.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getProductSpecialPrice}
+     *
      *
      * @param productId         the id of the product. Use it instead of productIdOrSku in
      *                          case you are sure the product identifier is a product id
@@ -1323,7 +1545,7 @@ public class MagentoCloudConnector {
                                                              @Optional String productSku,
                                                              @Optional String productIdOrSku,
                                                              @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().getProductSpecialPrice(ProductIdentifiers.from(productSku, productId, productIdOrSku),
+        return catalogClient.getProductSpecialPrice(ProductIdentifiers.from(productSku, productId, productIdOrSku),
                 storeViewIdOrCode);
     }
 
@@ -1333,6 +1555,7 @@ public class MagentoCloudConnector {
      * catalog-product-info SOAP method
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getProduct}
+     *
      *
      * @param productId                the id of the product. Use it instead of productIdOrSku in
      *                                 case you are sure the product identifier is a product id
@@ -1351,13 +1574,14 @@ public class MagentoCloudConnector {
                                                  @Optional String storeViewIdOrCode,
                                                  @Placement(group = "Standard Product Attributes to Retrieve") @Optional List<String> attributesNames,
                                                  @Placement(group = "Non-standard Product Attributes to Retrieve") @Optional List<String> additionalAttributeNames) {
-        return connection.getCatalogClient().getProduct(ProductIdentifiers.from(productSku, productId, productIdOrSku), storeViewIdOrCode, attributesNames, additionalAttributeNames);
+        return catalogClient.getProduct(ProductIdentifiers.from(productSku, productId, productIdOrSku), storeViewIdOrCode, attributesNames, additionalAttributeNames);
     }
 
     /**
      * Retrieve products list by filters. See catalog-product-list SOAP method.
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listProducts}
+     *
      *
      * @param filter            optional filtering expression - one or more comma-separated
      *                          unary or binary predicates, one for each filter, in the form
@@ -1371,7 +1595,7 @@ public class MagentoCloudConnector {
     @Processor
     public List<CatalogProductEntity> listProducts(@org.mule.api.annotations.Query @Optional String filter,
                                                    @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().listProducts(filter, storeViewIdOrCode);
+        return catalogClient.listProducts(filter, storeViewIdOrCode);
     }
 
 
@@ -1399,12 +1623,12 @@ public class MagentoCloudConnector {
                                          @Optional String fromDate,
                                          @Optional String toDate,
                                          @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().updateProductSpecialPrice(ProductIdentifiers.from(productSku, productId, productIdOrSku), specialPrice,
+        return catalogClient.updateProductSpecialPrice(ProductIdentifiers.from(productSku, productId, productIdOrSku), specialPrice,
                 fromDate, toDate, storeViewIdOrCode);
     }
 
     /**
-     * Updates a product. At least one of attributes or additionalAttributes
+     * Updates a product. At least one of attributes or additionalAttributes 
      * must be non null and non empty. See catalog-category-updateProduct SOAP method
      * <p/>
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateProduct}
@@ -1414,7 +1638,7 @@ public class MagentoCloudConnector {
      * @param productSku           the sku of the product. Use it instead of productIdOrSku in
      *                             case you are sure the product identifier is a product sku
      * @param productIdOrSku       the id or sku of the product.
-     * @param catalogProductEntity standard product attributes to update
+     * @param catalogProductEntity           standard product attributes to update
      * @param additionalAttributes non standard product attributes to update
      * @param storeViewIdOrCode    the id or code of the target store. Left unspecified for using current store
      * @return update result
@@ -1426,7 +1650,7 @@ public class MagentoCloudConnector {
                                  @Default("#[payload]") @Placement(group = "Standard Product Attributes to Update") CatalogProductCreateEntity catalogProductEntity,
                                  @Placement(group = "Non-standard Product Attributes to Update") @Optional List<AssociativeEntity> additionalAttributes,
                                  @Optional String storeViewIdOrCode) {
-        return connection.getCatalogClient().updateProduct(ProductIdentifiers.from(productSku, productId, productIdOrSku), catalogProductEntity, additionalAttributes, storeViewIdOrCode);
+        return catalogClient.updateProduct(ProductIdentifiers.from(productSku, productId, productIdOrSku), catalogProductEntity, additionalAttributes, storeViewIdOrCode);
     }
 
     /**
@@ -1438,12 +1662,13 @@ public class MagentoCloudConnector {
      */
     @Processor
     public int createShoppingCart(@Optional String storeId) {
-        return connection.getShoppingCartClient().createShoppingCart(storeId);
+        return shoppingCartClient.createShoppingCart(storeId);
     }
 
     /**
      * Retrieves full information about the shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:getInfoShoppingCart}
+     *
      *
      * @param quoteId Shopping cart ID (quote ID)
      * @param storeId Store view ID or code
@@ -1451,12 +1676,13 @@ public class MagentoCloudConnector {
      */
     @Processor
     public ShoppingCartInfoEntity getInfoShoppingCart(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().getShoppingCartInfo(quoteId, storeId);
+        return shoppingCartClient.getShoppingCartInfo(quoteId, storeId);
     }
 
     /**
      * Retrieves total prices for a shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listShoppingCartTotals}
+     *
      *
      * @param quoteId Shopping cart ID (quote identifier)
      * @param storeId Store view ID or code
@@ -1464,26 +1690,27 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<ShoppingCartTotalsEntity> listShoppingCartTotals(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().listShoppingCartTotals(quoteId, storeId);
+        return shoppingCartClient.listShoppingCartTotals(quoteId, storeId);
     }
 
     /**
      * Creates an order from a shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:createShoppingCartOrder}
      *
-     * @param quoteId  Shopping Cart ID (quote ID)
-     * @param storeId  Store view ID or code
+     * @param quoteId Shopping Cart ID (quote ID)
+     * @param storeId Store view ID or code
      * @param licenses Website license ID
      * @return result code of order creation
      */
     @Processor
     public String createShoppingCartOrder(int quoteId, @Optional String storeId, @Placement(group = "Licenses") @Optional List<String> licenses) {
-        return connection.getShoppingCartClient().createShoppingCartOrder(quoteId, storeId, licenses);
+        return shoppingCartClient.createShoppingCartOrder(quoteId, storeId, licenses);
     }
 
     /**
      * Retrieves the website license agreement for the quote according to the website (store). Store id must be given in cart creation.
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listShoppingCartLicenses}
+     *
      *
      * @param quoteId Shopping Cart ID (quote ID)
      * @param storeId Store view ID or code
@@ -1491,60 +1718,61 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<ShoppingCartLicenseEntity> listShoppingCartLicenses(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().listShoppingCartLicenses(quoteId, storeId);
+        return shoppingCartClient.listShoppingCartLicenses(quoteId, storeId);
     }
 
     /**
      * Adds one or more products to the shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:addShoppingCartProduct}
      *
-     * @param quoteId  Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param products Products data
-     * @param storeId  Store view ID or code
+     * @param storeId Store view ID or code
      * @return True on success (if the product is added to the shopping cart)
      */
     @Processor
     public boolean addShoppingCartProduct(int quoteId,
                                           @Default("#[payload]") @Placement(group = "Products attributes") List<ShoppingCartProductEntity> products,
                                           @Optional String storeId) {
-        return connection.getShoppingCartClient().addShoppingCartProduct(quoteId, products, storeId);
+        return shoppingCartClient.addShoppingCartProduct(quoteId, products, storeId);
     }
 
     /**
      * Updates one or several products in the shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:updateShoppingCartProduct}
      *
-     * @param quoteId  Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param products Products data
-     * @param storeId  Store view ID or code
+     * @param storeId Store view ID or code
      * @return True if the product is updated
      */
     @Processor
     public boolean updateShoppingCartProduct(int quoteId,
                                              @Default("#[payload]") @Placement(group = "Products attributes") List<ShoppingCartProductEntity> products,
                                              @Optional String storeId) {
-        return connection.getShoppingCartClient().updateShoppingCartProduct(quoteId, products, storeId);
+        return shoppingCartClient.updateShoppingCartProduct(quoteId, products, storeId);
     }
 
     /**
      * Removes one or several products from a shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:removeShoppingCartProduct}
      *
-     * @param quoteId  Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param products Products data
-     * @param storeId  Store view ID or code
+     * @param storeId Store view ID or code
      * @return True if the product is removed
      */
     @Processor
     public boolean removeShoppingCartProduct(int quoteId,
                                              @Default("#[payload]") @Placement(group = "Products attributes") List<ShoppingCartProductEntity> products,
                                              @Optional String storeId) {
-        return connection.getShoppingCartClient().removeShoppingCartProduct(quoteId, products, storeId);
+        return shoppingCartClient.removeShoppingCartProduct(quoteId, products, storeId);
     }
 
     /**
      * Retrieves the list of products in the shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listShoppingCartProducts}
+     *
      *
      * @param quoteId Shopping Cart ID (quote ID)
      * @param storeId Store view ID or code
@@ -1552,55 +1780,55 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<CatalogProductEntity> listShoppingCartProducts(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().listShoppingCartProducts(quoteId, storeId);
+        return shoppingCartClient.listShoppingCartProducts(quoteId, storeId);
     }
 
     /**
      * Moves products from the current quote to a customer quote.
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:moveShoppingCartProductToCustomerQuote}
      *
-     * @param quoteId  Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param products Products data
-     * @param storeId  Store view ID or code
+     * @param storeId Store view ID or code
      * @return True if the product is moved to customer quote
      */
     @Processor
     public boolean moveShoppingCartProductToCustomerQuote(int quoteId,
                                                           @Default("#[payload]") @Placement(group = "Products attributes") List<ShoppingCartProductEntity> products,
                                                           @Optional String storeId) {
-        return connection.getShoppingCartClient().moveShoppingCartProductToCustomerQuote(quoteId, products, storeId);
+        return shoppingCartClient.moveShoppingCartProductToCustomerQuote(quoteId, products, storeId);
     }
 
     /**
      * Adds information about the customer to a shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:setShoppingCartCustomer}
      *
-     * @param quoteId  Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param customer Cuestomer data
-     * @param storeId  Store view ID or code
+     * @param storeId Store view ID or code
      * @return True if information is added
      */
     @Processor
     public boolean setShoppingCartCustomer(int quoteId,
                                            @Default("#[payload]") @Placement(group = "Customer attributes") ShoppingCartCustomerEntity customer,
                                            @Optional String storeId) {
-        return connection.getShoppingCartClient().setShoppingCartCustomer(quoteId, customer, storeId);
+        return shoppingCartClient.setShoppingCartCustomer(quoteId, customer, storeId);
     }
 
     /**
      * Sets the customer addresses in the shopping cart (quote). Don't set a coupon to the cart before using this method, the address won't be set.
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:setShoppingCartCustomerAddresses}
      *
-     * @param quoteId                       Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param shoppingCartCustomerAddresses Addresses data (mode is required for each address and should be either "shipping" or "billing")
-     * @param storeId                       Store view ID or code
+     * @param storeId Store view ID or code
      * @return True if information is added
      */
     @Processor
     public boolean setShoppingCartCustomerAddresses(int quoteId,
                                                     @Default("#[payload]") @Placement(group = "Addresses attributes") List<ShoppingCartCustomerAddressEntity> shoppingCartCustomerAddresses,
                                                     @Optional String storeId) {
-        return connection.getShoppingCartClient().setShoppingCartCustomerAddresses(quoteId, shoppingCartCustomerAddresses, storeId);
+        return shoppingCartClient.setShoppingCartCustomerAddresses(quoteId, shoppingCartCustomerAddresses, storeId);
     }
 
     /**
@@ -1608,18 +1836,19 @@ public class MagentoCloudConnector {
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:setShoppingCartShippingMethod}
      *
      * @param quoteId Shopping Cart ID (quote ID)
-     * @param method  Shipping method code
+     * @param method Shipping method code
      * @param storeId Store view ID or code
      * @return True if the shipping method is retrieved
      */
     @Processor
     public boolean setShoppingCartShippingMethod(int quoteId, String method, @Optional String storeId) {
-        return connection.getShoppingCartClient().setShoppingCartShippingMethod(quoteId, method, storeId);
+        return shoppingCartClient.setShoppingCartShippingMethod(quoteId, method, storeId);
     }
 
     /**
      * Retrieves the list of available shipping methods for a shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listShoppingCartShippingMethods}
+     *
      *
      * @param quoteId Shopping Cart ID (quote ID)
      * @param storeId Store view ID or code
@@ -1627,28 +1856,29 @@ public class MagentoCloudConnector {
      */
     @Processor
     public List<ShoppingCartShippingMethodEntity> listShoppingCartShippingMethods(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().listShoppingCartShippingMethods(quoteId, storeId);
+        return shoppingCartClient.listShoppingCartShippingMethods(quoteId, storeId);
     }
 
     /**
      * Sets a payment method for a shopping cart (quote), cart must not be empty.
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:setShoppingCartPaymentMethod}
      *
-     * @param quoteId                   Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param shoppingCartPaymentMethod Payment mehthod data
-     * @param storeId                   Store view ID or code
+     * @param storeId Store view ID or code
      * @return True on success
      */
     @Processor
     public boolean setShoppingCartPaymentMethod(int quoteId,
                                                 @Default("#[payload]") @Placement(group = "Method attributes") ShoppingCartPaymentMethodEntity shoppingCartPaymentMethod,
                                                 @Optional String storeId) {
-        return connection.getShoppingCartClient().setShoppingCartPaymentMethod(quoteId, shoppingCartPaymentMethod, storeId);
+        return shoppingCartClient.setShoppingCartPaymentMethod(quoteId, shoppingCartPaymentMethod, storeId);
     }
 
     /**
      * Retrieves a list of available payment methods for a shopping cart (quote).
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:listShoppingCartPaymentMethods}
+     *
      *
      * @param quoteId Shopping Cart ID (quote ID)
      * @param storeId Store view ID or code
@@ -1656,21 +1886,21 @@ public class MagentoCloudConnector {
      */
     @Processor
     public ShoppingCartPaymentMethodResponseEntityArray listShoppingCartPaymentMethods(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().listShoppingCartPaymentMethods(quoteId, storeId);
+        return shoppingCartClient.listShoppingCartPaymentMethods(quoteId, storeId);
     }
 
     /**
      * Adds a coupon code for a shopping cart (quote). The shopping cart must not be empty. The shopping cart must be created with a storeId for this operation to work.
      * {@sample.xml ../../../doc/magento-connector.xml.sample magento:addShoppingCartCoupon}
      *
-     * @param quoteId    Shopping Cart ID (quote ID)
+     * @param quoteId Shopping Cart ID (quote ID)
      * @param couponCode Coupon code
-     * @param storeId    Store view ID or code
+     * @param storeId Store view ID or code
      * @return True if the coupon code is added
      */
     @Processor
     public boolean addShoppingCartCoupon(int quoteId, String couponCode, @Optional String storeId) {
-        return connection.getShoppingCartClient().addShoppingCartCoupon(quoteId, couponCode, storeId);
+        return shoppingCartClient.addShoppingCartCoupon(quoteId, couponCode, storeId);
     }
 
     /**
@@ -1683,6 +1913,72 @@ public class MagentoCloudConnector {
      */
     @Processor
     public boolean removeShoppingCartCoupon(int quoteId, @Optional String storeId) {
-        return connection.getShoppingCartClient().removeShoppingCartCoupon(quoteId, storeId);
+        return shoppingCartClient.removeShoppingCartCoupon(quoteId, storeId);
+    }
+
+    @SuppressWarnings("unchecked")
+    public void setOrderClient(MagentoOrderClient<?> magentoOrderClient) {
+        this.orderClient = MagentoClientAdaptor.adapt(MagentoOrderClient.class, magentoOrderClient);
+    }
+
+    @SuppressWarnings("unchecked")
+    public void setCustomerClient(MagentoCustomerClient<?> customerClient) {
+        this.customerClient = MagentoClientAdaptor.adapt(MagentoCustomerClient.class, customerClient);
+    }
+
+    @SuppressWarnings("unchecked")
+    public void setInventoryClient(MagentoInventoryClient<?> inventoryClient) {
+        this.inventoryClient = MagentoClientAdaptor.adapt(MagentoInventoryClient.class, inventoryClient);
+    }
+
+    @SuppressWarnings("unchecked")
+    public void setDirectoryClient(MagentoDirectoryClient<?> directoryClient) {
+        this.directoryClient = MagentoClientAdaptor.adapt(MagentoDirectoryClient.class, directoryClient);
+    }
+
+    @SuppressWarnings("unchecked")
+    public void setCatalogClient(MagentoCatalogClient<?> catalogClient) {
+        this.catalogClient = MagentoClientAdaptor.adapt(MagentoCatalogClient.class, catalogClient);
+    }
+
+    @SuppressWarnings("unchecked")
+    public void setShoppingCartClient(MagentoShoppingCartClient<?> catalogClient) {
+        this.shoppingCartClient = MagentoClientAdaptor.adapt(MagentoShoppingCartClient.class, catalogClient);
+    }
+
+    public String getUser() {
+        return user;
+    }
+
+    public void setUser(String user) {
+        this.user = user;
+    }
+
+    public String getServerAddress() {
+        return serverAddress;
+    }
+
+    public void setServerAddress(String serverAddress) {
+        this.serverAddress = serverAddress;
+    }
+
+    private class PortProviderInitializer {
+        private DefaultAxisPortProvider provider;
+        private String username;
+        private String password;
+        private String address;
+
+        public PortProviderInitializer(String username, String password, String address) {
+            this.username = username;
+            this.password = password;
+            this.address = address;
+        }
+
+        public AxisPortProvider getPortProvider() {
+            if (provider == null) {
+                provider = new DefaultAxisPortProvider(username, password, address);
+            }
+            return provider;
+        }
     }
 }
